calculate_unit_precision function implementation and its test Which calculates a coefficient to do - Models::Unit float->int coversion - and balancing between different unit capacities
New test file for tests related to Filter class
+ test against Integer Range Error: Too big to convert
+ test to check if run_ortools gets the correct unit.precision_coef's

TODO: Implementation of the function is done but the precision_coef is
not in use. Need to replace "*1000" multiplications with
"unit.precision_coef". Warning: overload_multiplier needs to be
divided by this coeff to keep the problem unchanged.
(And potentially other cost coefs related to quantity and capacity.)